fix #1: Add simple input validation to server name, ip, and port fields
This also adds logging to a file, along with the console.

new server == 201/Created
updating server == 200/OK

The logic for determining if we have seen the server is just using IP for now, but should use IP and port together since that's the smallest set of information you could use to describe a network endpoint.

Bonus: If you supply the -p option with a port number, it will bind to a TCP port of your choosing, not the default of TCP/8080.