Fix #3: Add /getip endpoint for clients to nab their Internet-facing IPv4 address
This also adds a command line option -a so you can specify the local IP to bind to, with a default to "everything" 0.0.0.0. Some local systems will return v6 link local addresses as the source IP here and you should be able to force v4. If you still want to listen on localhost, be explicit and use 127.0.0.1.