Fix #2: Test for membership in servers map instead of iterating, add cli arg for staleThreshold, reduce scope of thresholds
I brought the stale and calculated prune thresholds into the closest scope were the pruneServers function needed them. There was no need for them to be global.

mux lock/unlock is only called once, return is removed, and we give different responses to the client in a concise if/else block.